Small changes to containers and tests.

+++ src/util/misc/Memory.cpp
@@ -14,10 +14,10 @@ namespace Util
 {
 
    /// Total amount of memory allocated, in bytes.
-   size_t Memory::total_ = 0;
+   int Memory::total_ = 0;
 
    /// Maximum of total over course of simulation.
-   size_t Memory::max_ = 0;
+   int Memory::max_ = 0;
 
    /*
    * Call this to ensure compilation of this file. 
@@ -31,25 +31,25 @@ namespace Util
    /*
    * Return total amount of memory allocated thus far.
    */
-   size_t Memory::total()
+   int Memory::total()
    { return total_; }
 
    /*
    * Return maximum amount of allocated memory thus far.
    */
-   size_t Memory::max()
+   int Memory::max()
    { return max_; }
 
    /*
    * Compute memory usage statistics (call on all processors).
    */
    #ifdef UTIL_MPI
-   size_t Memory::max(MPI::Intracomm& communicator)
+   int Memory::max(MPI::Intracomm& communicator)
    { 
       int maxGlobal;
       int maxLocal = max_;
       communicator.Allreduce(&maxLocal, &maxGlobal, 1, MPI::INT, MPI::MAX);
-      return (size_t)maxGlobal;
+      return maxGlobal;
    }
    #endif
